Support bootloader v9.1 .
The bootloader of the STM32H753II was updated to v9.1 . The new
version disables the clock for the USART3 peripheral. As the stub
uses USART3 for comminucation, the clock must be enabled again.
Please note that this patch is compatible with the bootloader v9.0
where the clock is already enabled.